+++ package.json
@@ -12,6 +12,7 @@
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.1",
     "react-notifications": "^1.7.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "typescript": "^4.2.3",
     "web-vitals": "^1.1.0"
@@ -42,6 +43,7 @@
   },
   "devDependencies": {
     "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.2"
+    "@types/react-dom": "^17.0.2",
+    "@types/react-router-dom": "^5.1.7"
   }
 }
