+++ src/containers/Weather/Weather.tsx
@@ -1,0 +1,7 @@
+export const Weather = () => {
+  return (
+    <div>
+      <p>This is the weather page</p>
+    </div>
+  );
+};
